+++ CHANGELOG.md
@@ -11,6 +11,7 @@ I know that many can't update to the latest build. The package will continue wor
 * Added: extended boolean scope to constant.language.boolean
 * Added: extended magic constant scope to constant.language.magic
 * Added: extended null constant scope to "constant.language.null"
+* Added: punctuation definition scope to the parens of type casts
 * Fixed: #56 indentation for empty brackets
 * Fixed: #58 syntax bracket handling for closures
 * Fixed: #61, #62, and #63 operator scopes not matching correctly